Fix cropping CLI
As observed by @bertsky, the latest changes to the code disabled
the cropping functionality. Since DFKI cropping is the only
dedicated processor for this task in OCR-D, we need it asap. This
commit proposes the necessary changes to bring cropping back to
work. However, it is far from optimal. I.e. the heavy dependencies
introduced by the machine-learnign based tools in this repo should
be somehow circumvented for the relatively simple rule-based
cropping processor.

Fixes #11.